# CUDA: mul_mat_v support for batch sizes > 1