@IMbackK 2025-06-18
(hip)rocblas performing very poorly on rdna is a known issue and not down to the exact calls we are us, its pretty bad for rdna2 but it gets worse for rdna3 and for rnda4 it might as well be broken performance wise.

On mi hardware the performance is much better so possibly we will not want to do this there, but needs bench marking.