@JohannesGaessler 2025-06-18
I forgot: I changed the integer size in the kernel from 64 bit to 32 bit due to issues with register pressure.