@JohannesGaessler 2025-06-18
I changed specifically the calculation of the initial offsets to 64 bit math. That is the only part of the kernel where the byte offsets scale with the product of 2 tensor dimensions. The byte offsets scaling with 1 tensor dimension are at least 1024x lower.