@JohannesGaessler 2025-06-18
Thank you, I forgot to check `MUL_MAT_ID` for the final version.