@IMbackK 2025-06-19
On cdna i am seeing a large (2x +) slow down starting at batch 4 in all datatypes.