@IMbackK 2025-06-22
btw expiramenting with rocblas-bench shows we would be better off using rocblas_hshgemv_batched and friends.